Just checking in / update
I was thinking of a new syntax like so:

wave(value)
  .match(branch)
  .match(branch)
  .match(branch)

But the issue is how do we evaluate/call/unwrap and get the value?  The only
thing returned would be some object with a .match function but not an actual
value. Then we would have to create some lame API like
wave(value).match(branch).getValue() or something like that. Skip, for now.